[Finish]Laravel入門編4

+++ backend/resources/views/show.blade.php
@@ -16,11 +16,15 @@
 
   <div>
     <a href={{ route("shop.list") }}>一覧に戻る</a>
-    | <a href={{ route("shop.edit", ["id" => $shop->id]) }}>編集</a>
-    <p>
-      {{ Form::open(["method" => "delete", "route" => ["shop.destroy", $shop->id]]) }}
-        {{ Form::submit("削除", ["class" => "btn btn-outline-danger"]) }}
-      {{ Form::close() }}
-    </p>
+    @auth
+      @if ($shop->user_id === $login_user_id)
+          | <a href={{ route("shop.edit", ["id" => $shop->id]) }}>編集</a>
+        <p>
+          {{ Form::open(["method" => "delete", "route" => ["shop.destroy", $shop->id]]) }}
+            {{ Form::submit("削除", ["class" => "btn btn-outline-danger"]) }}
+          {{ Form::close() }}
+        </p>
+      @endif
+    @endauth
   </div>
 @endsection
